+++ src/main/java/backend/siptis/service/userData/SiptisUserServiceImpl.java
@@ -16,7 +16,7 @@ public class SiptisUserServiceImpl implements SiptisUserService {
     private final SiptisUserRepository usuarioCommonRepository;
 
     @Override
-    public ServiceAnswer obtenerProyectosSupervisorParaMenuPrincipalPorIdUsuario(Integer id) {
+    public ServiceAnswer obtenerProyectosSupervisorParaMenuPrincipalPorIdUsuario(Long id) {
         Optional<SiptisUser> usuarioOptional = usuarioCommonRepository.findById(id);
         if(usuarioOptional.isEmpty()){
             return ServiceAnswer.builder().serviceMessage(ServiceMessage.NOT_FOUND).data(null).build();
